fix(build): mlkit 0.13.x String API + clamp toDouble() + Flutter 3.27
google_mlkit_translation 0.13.x: TranslateLanguage → String 코드
- isModelDownloaded/downloadModel 인자: TranslateLanguage → 'en','ko','ja','zh'
- OnDeviceTranslator 생성자: enum → String 코드
- pubspec: ^0.12.0 → ^0.13.0

camera_overlay: clamp() 리턴 num → .toDouble() 명시

+++ apps/parksy-studio/lib/widgets/camera_overlay.dart
@@ -29,15 +29,15 @@ class _CameraOverlayState extends State<CameraOverlay> {
     return LayoutBuilder(
       builder: (context, constraints) {
         // 절대 좌표 clamp: 부모 크기 - 오버레이 크기
-        final maxDx = constraints.maxWidth.isFinite  ? constraints.maxWidth  / 2 : 200;
-        final maxDy = constraints.maxHeight.isFinite ? constraints.maxHeight / 2 : 400;
+        final maxDx = constraints.maxWidth.isFinite  ? constraints.maxWidth  / 2 : 200.0;
+        final maxDy = constraints.maxHeight.isFinite ? constraints.maxHeight / 2 : 400.0;
 
         return GestureDetector(
           onPanUpdate: (d) {
             setState(() {
               _offset = Offset(
-                (_offset.dx + d.delta.dx).clamp(-maxDx, maxDx),
-                (_offset.dy + d.delta.dy).clamp(-maxDy, maxDy),
+                (_offset.dx + d.delta.dx).clamp(-maxDx, maxDx).toDouble(),
+                (_offset.dy + d.delta.dy).clamp(-maxDy, maxDy).toDouble(),
               );
             });
           },
